test: split out shared tests into web and native
