+++ packages/article/__tests__/images-with-style.native.js
@@ -6,7 +6,7 @@ import {
   print
 } from "@times-components/jest-serializer";
 import "./mocks.native";
-import shared from "./images.base";
+import shared from "./images.base.native";
 
 jest.mock("../src/article-comments/article-comments", () => "ArticleComments");
